djangobackend/requirements.txt
django
mysqlclient
django-cors-headers
django-rest-framework
django-extensions
redis
pytz
requests
djangorestframework_simplejwt


